feat: centralize page callbacks
