+++ yosai_intel_dashboard/src/adapters/ui/pages/greetings/__init__.py
@@ -1,4 +1,4 @@
-from .callbacks import register_callbacks
 from .layout import layout
+from yosai_intel_dashboard.src.callbacks.controller import register_greetings_callbacks as register_callbacks
 
 __all__ = ["layout", "register_callbacks"]
