+++ yosai_intel_dashboard/src/app.py
@@ -2,9 +2,17 @@
 
 from dash import Dash, html
 
-from yosai_intel_dashboard.src.pages import greetings, register_callbacks
-from yosai_intel_dashboard.src.services.greeting import GreetingService
+from yosai_intel_dashboard.src.adapters.ui.pages import greetings
+from yosai_intel_dashboard.src.callbacks import register_callbacks
 from yosai_intel_dashboard.src.simple_di import ServiceContainer
+
+try:  # pragma: no cover - allow running without full services package
+    from yosai_intel_dashboard.src.services.greeting import GreetingService
+except Exception:  # pragma: no cover - lightweight fallback
+
+    class GreetingService:  # type: ignore[too-few-public-methods]
+        def greet(self, name: str) -> str:
+            return f"Hello, {name}!"
 
 
 def create_app() -> Dash:
